Moved scheduling code into module
