+++ src/test/unit/scheduler/test_analysis.py
@@ -221,115 +221,6 @@ class TestAnalysisSchedulerBlacklist:
         self.sched.config.set('test_plugin', 'mime_blacklist', 'type1, type2')
 
 
-class TestUtilityFunctions:
-
-    class PluginMock:
-        def __init__(self, dependencies):
-            self.DEPENDENCIES = dependencies
-
-    @classmethod
-    def setup_class(cls):
-        cls.init_patch = mock.patch(target='scheduler.analysis.AnalysisScheduler.__init__', new=lambda *_: None)
-        cls.init_patch.start()
-        cls.scheduler = AnalysisScheduler()
-        cls.plugin_list = ['no_deps', 'foo', 'bar']
-        cls.init_patch.stop()
-
-    def _add_plugins(self):
-        self.scheduler.analysis_plugins = {
-            'no_deps': self.PluginMock(dependencies=[]),
-            'foo': self.PluginMock(dependencies=['no_deps']),
-            'bar': self.PluginMock(dependencies=['no_deps', 'foo'])
-        }
-
-    def _add_plugins_with_recursive_dependencies(self):
-        self.scheduler.analysis_plugins = {
-            'p1': self.PluginMock(['p2', 'p3']),
-            'p2': self.PluginMock(['p3']),
-            'p3': self.PluginMock([]),
-            'p4': self.PluginMock(['p5']),
-            'p5': self.PluginMock(['p6']),
-            'p6': self.PluginMock([])
-        }
-
-    @pytest.mark.parametrize('input_data, expected_output', [
-        (set(), set()),
-        ({'p1'}, {'p2', 'p3'}),
-        ({'p3'}, set()),
-        ({'p1', 'p2', 'p3', 'p4'}, {'p5'}),
-    ])
-    def test_get_cumulative_remaining_dependencies(self, input_data, expected_output):
-        self._add_plugins_with_recursive_dependencies()
-        result = self.scheduler._get_cumulative_remaining_dependencies(input_data)
-        assert result == expected_output
-
-    @pytest.mark.parametrize('input_data, expected_output', [
-        ([], set()),
-        (['p3'], {'p3'}),
-        (['p1'], {'p1', 'p2', 'p3'}),
-        (['p4'], {'p4', 'p5', 'p6'}),
-    ])
-    def test_add_dependencies_recursively(self, input_data, expected_output):
-        self._add_plugins_with_recursive_dependencies()
-        result = self.scheduler._add_dependencies_recursively(input_data)
-        assert set(result) == expected_output
-
-    @pytest.mark.parametrize('remaining, scheduled, expected_output', [
-        ({}, [], []),
-        ({'no_deps', 'foo', 'bar'}, [], ['no_deps']),
-        ({'foo', 'bar'}, ['no_deps'], ['foo']),
-        ({'bar'}, ['no_deps', 'foo'], ['bar']),
-    ])
-    def test_get_plugins_with_met_dependencies(self, remaining, scheduled, expected_output):
-        self._add_plugins()
-        assert self.scheduler._get_plugins_with_met_dependencies(remaining, scheduled) == expected_output
-
-    @pytest.mark.parametrize('remaining, scheduled, expected_output', [
-        ({'bar'}, ['no_deps', 'foo'], {'bar'}),
-        ({'foo', 'bar'}, ['no_deps', 'foo'], {'foo', 'bar'}),
-    ])
-    def test_get_plugins_with_met_dependencies__completed_analyses(self, remaining, scheduled, expected_output):
-        self._add_plugins()
-        assert set(self.scheduler._get_plugins_with_met_dependencies(remaining, scheduled)) == expected_output
-
-    def test_reschedule_failed_analysis_task(self):
-        task = Firmware(binary='foo')
-        error_message = 'There was an exception'
-        task.analysis_exception = ('foo', error_message)
-        task.scheduled_analysis = ['no_deps', 'bar']
-        task.processed_analysis['foo'] = {'error': 1}
-        self._add_plugins()
-        self.scheduler._reschedule_failed_analysis_task(task)
-
-        assert 'foo' in task.processed_analysis
-        assert task.processed_analysis['foo'] == {'failed': error_message}
-        assert 'bar' not in task.scheduled_analysis
-        assert 'bar' in task.processed_analysis
-        assert task.processed_analysis['bar'] == {'failed': 'Analysis of dependency foo failed'}
-        assert 'no_deps' in task.scheduled_analysis
-
-    def test_smart_shuffle(self):
-        self._add_plugins()
-        result = self.scheduler._smart_shuffle(self.plugin_list)
-        assert result == ['bar', 'foo', 'no_deps']
-
-    def test_smart_shuffle__impossible_dependency(self):
-        self._add_plugins()
-        self.scheduler.analysis_plugins['impossible'] = self.PluginMock(dependencies=['impossible to meet'])
-        result = self.scheduler._smart_shuffle(self.plugin_list + ['impossible'])
-        assert 'impossible' not in result
-        assert result == ['bar', 'foo', 'no_deps']
-
-    def test_smart_shuffle__circle_dependency(self):
-        self.scheduler.analysis_plugins = {
-            'p1': self.PluginMock(['p2']),
-            'p2': self.PluginMock(['p3']),
-            'p3': self.PluginMock(['p1']),
-        }
-        result = self.scheduler._smart_shuffle(['p1', 'p2', 'p3'])
-        assert result == []
-
-
 class TestAnalysisSkipping:
 
     class PluginMock:
